Use SszVectorSchema, SszListSchema, SszCollectionSchema interfaces, hide implementation classes. Refactor RPC message schemas to extends AbstractSszListSchema directly
