+++ ethereum/datastructures/src/main/java/tech/pegasys/teku/datastructures/blocks/BeaconBlockBody.java
@@ -70,37 +70,37 @@ public class BeaconBlockBody
           namedSchema("graffiti", SszPrimitiveSchemas.BYTES32_SCHEMA),
           namedSchema(
               "proposer_slashings",
-              new SszListSchema<>(ProposerSlashing.SSZ_SCHEMA, Constants.MAX_PROPOSER_SLASHINGS)),
+              SszListSchema.create(ProposerSlashing.SSZ_SCHEMA, Constants.MAX_PROPOSER_SLASHINGS)),
           namedSchema(
               "attester_slashings",
-              new SszListSchema<>(AttesterSlashing.SSZ_SCHEMA, Constants.MAX_ATTESTER_SLASHINGS)),
+              SszListSchema.create(AttesterSlashing.SSZ_SCHEMA, Constants.MAX_ATTESTER_SLASHINGS)),
           namedSchema(
               "attestations",
-              new SszListSchema<>(Attestation.SSZ_SCHEMA, Constants.MAX_ATTESTATIONS)),
-          namedSchema("deposits", new SszListSchema<>(Deposit.SSZ_SCHEMA, Constants.MAX_DEPOSITS)),
+              SszListSchema.create(Attestation.SSZ_SCHEMA, Constants.MAX_ATTESTATIONS)),
+          namedSchema("deposits", SszListSchema.create(Deposit.SSZ_SCHEMA, Constants.MAX_DEPOSITS)),
           namedSchema(
               "voluntary_exits",
-              new SszListSchema<>(SignedVoluntaryExit.SSZ_SCHEMA, Constants.MAX_VOLUNTARY_EXITS)));
+              SszListSchema.create(SignedVoluntaryExit.SSZ_SCHEMA, Constants.MAX_VOLUNTARY_EXITS)));
     }
 
-    public SszListSchema<ProposerSlashing> getProposerSlashingsSchema() {
-      return (SszListSchema<ProposerSlashing>) getFieldSchema3();
+    public SszListSchema<ProposerSlashing, ?> getProposerSlashingsSchema() {
+      return (SszListSchema<ProposerSlashing, ?>) getFieldSchema3();
     }
 
-    public SszListSchema<AttesterSlashing> getAttesterSlashingsSchema() {
-      return (SszListSchema<AttesterSlashing>) getFieldSchema4();
+    public SszListSchema<AttesterSlashing, ?> getAttesterSlashingsSchema() {
+      return (SszListSchema<AttesterSlashing, ?>) getFieldSchema4();
     }
 
-    public SszListSchema<Attestation> getAttestationsSchema() {
-      return (SszListSchema<Attestation>) getFieldSchema5();
+    public SszListSchema<Attestation, ?> getAttestationsSchema() {
+      return (SszListSchema<Attestation, ?>) getFieldSchema5();
     }
 
-    public SszListSchema<Deposit> getDepositsSchema() {
-      return (SszListSchema<Deposit>) getFieldSchema6();
+    public SszListSchema<Deposit, ?> getDepositsSchema() {
+      return (SszListSchema<Deposit, ?>) getFieldSchema6();
     }
 
-    public SszListSchema<SignedVoluntaryExit> getVoluntaryExitsSchema() {
-      return (SszListSchema<SignedVoluntaryExit>) getFieldSchema7();
+    public SszListSchema<SignedVoluntaryExit, ?> getVoluntaryExitsSchema() {
+      return (SszListSchema<SignedVoluntaryExit, ?>) getFieldSchema7();
     }
 
     @Override
