+++ ethereum/datastructures/src/main/java/tech/pegasys/teku/datastructures/operations/Deposit.java
@@ -37,7 +37,7 @@ public class Deposit extends Container2<Deposit, SszVector<SszBytes32>, DepositD
           "Deposit",
           namedSchema(
               "proof",
-              new SszVectorSchema<>(
+              SszVectorSchema.create(
                   SszPrimitiveSchemas.BYTES32_SCHEMA, Constants.DEPOSIT_CONTRACT_TREE_DEPTH + 1)),
           namedSchema("data", DepositData.SSZ_SCHEMA));
     }
