--- ssz/src/main/java/tech/pegasys/teku/ssz/backing/schema/SszCollectionSchemaIfc.java
@@ -1,11 +1,0 @@
-package tech.pegasys.teku.ssz.backing.schema;
-
-import tech.pegasys.teku.ssz.backing.SszCollection;
-import tech.pegasys.teku.ssz.backing.SszData;
-
-public interface SszCollectionSchemaIfc<
-    SszElementT extends SszData, SszCollectionT extends SszCollection<SszElementT>>
-    extends SszCompositeSchema<SszCollectionT> {
-
-  SszSchema<SszElementT> getElementSchema();
-}
