+++ ssz/src/main/java/tech/pegasys/teku/ssz/backing/view/SszMutableVectorImpl.java
@@ -16,7 +16,6 @@ package tech.pegasys.teku.ssz.backing.view;
 import tech.pegasys.teku.ssz.backing.SszData;
 import tech.pegasys.teku.ssz.backing.SszMutableRefVector;
 import tech.pegasys.teku.ssz.backing.SszMutableVector;
-import tech.pegasys.teku.ssz.backing.SszVector;
 import tech.pegasys.teku.ssz.backing.cache.IntCache;
 import tech.pegasys.teku.ssz.backing.schema.SszVectorSchema;
 import tech.pegasys.teku.ssz.backing.tree.TreeNode;
