+++ ssz/src/test/java/tech/pegasys/teku/ssz/backing/ListViewHintsTest.java
@@ -43,7 +43,7 @@ import tech.pegasys.teku.ssz.sos.SszReader;
 public class ListViewHintsTest {
 
   <TElement extends SszData> List<SszList<TElement>> createListVariants(
-      SszListSchema<TElement> type, SszList<TElement> list0) {
+      SszListSchema<TElement, ?> type, SszList<TElement> list0) {
     List<SszList<TElement>> ret = new ArrayList<>();
     ret.add(list0);
     if (!(list0 instanceof SszMutableData)) {
@@ -54,12 +54,12 @@ public class ListViewHintsTest {
   }
 
   <TElement extends SszData> void assertEmptyListVariants(
-      SszListSchema<TElement> type, SszList<TElement> list0) {
+      SszListSchema<TElement, ?> type, SszList<TElement> list0) {
     createListVariants(type, list0).forEach(l -> assertEmptyList(type, l));
   }
 
   <TElement extends SszData> void assertEmptyList(
-      SszListSchema<TElement> type, SszList<TElement> list) {
+      SszListSchema<TElement, ?> type, SszList<TElement> list) {
 
     if (!(list instanceof SszMutableData)) {
       assertThat(list.hashTreeRoot()).isEqualTo(type.getDefaultTree().hashTreeRoot());
@@ -74,12 +74,12 @@ public class ListViewHintsTest {
   }
 
   <TElement extends SszData> void assertListElementsVariants(
-      SszListSchema<TElement> type, SszList<TElement> list0, List<TElement> expectedElements) {
+      SszListSchema<TElement, ?> type, SszList<TElement> list0, List<TElement> expectedElements) {
     createListVariants(type, list0).forEach(l -> assertListElements(type, l, expectedElements));
   }
 
   <TElement extends SszData> void assertListElements(
-      SszListSchema<TElement> type, SszList<TElement> list, List<TElement> expectedElements) {
+      SszListSchema<TElement, ?> type, SszList<TElement> list, List<TElement> expectedElements) {
 
     assertThat(list.isEmpty()).isFalse();
     assertThat(list.size()).isEqualTo(expectedElements.size());
@@ -93,7 +93,7 @@ public class ListViewHintsTest {
   }
 
   <TElement extends SszData> void assertListEqualsVariants(
-      SszListSchema<TElement> type, SszList<TElement> list1, SszList<TElement> list2) {
+      SszListSchema<TElement, ?> type, SszList<TElement> list1, SszList<TElement> list2) {
     List<SszList<TElement>> listVariants1 = createListVariants(type, list1);
     List<SszList<TElement>> listVariants2 = createListVariants(type, list2);
 
@@ -104,7 +104,7 @@ public class ListViewHintsTest {
   }
 
   <TElement extends SszData> void assertListEquals(
-      SszListSchema<TElement> type, SszList<TElement> list1, SszList<TElement> list2) {
+      SszListSchema<TElement, ?> type, SszList<TElement> list1, SszList<TElement> list2) {
 
     assertThat(list1.size()).isEqualTo(list2.size());
     assertThat(list1).isEqualTo(list2);
@@ -179,8 +179,8 @@ public class ListViewHintsTest {
         Arguments.of(listElementType4, 17, elementSupplier4));
   }
 
-  static <TElement extends SszData> List<SszListSchema<TElement>> generateTypesWithHints(
-      SszListSchema<TElement> originalType) {
+  static <TElement extends SszData> List<SszListSchema<TElement, ?>> generateTypesWithHints(
+      SszListSchema<TElement, ?> originalType) {
     return Stream.concat(
             Stream.of(originalType),
             IntStream.of(0, 1, 2, 4, 8, 10)
@@ -188,7 +188,7 @@ public class ListViewHintsTest {
                 .mapToObj(SszSchemaHints::sszSuperNode)
                 .map(
                     typeHints ->
-                        new SszListSchema<TElement>(
+                        SszListSchema.create(
                             originalType.getElementSchema(),
                             originalType.getMaxLength(),
                             typeHints)))
@@ -202,15 +202,15 @@ public class ListViewHintsTest {
       long maxListSize,
       Supplier<TElement> listElementsFactory) {
 
-    List<SszListSchema<TElement>> types =
-        generateTypesWithHints(new SszListSchema<>(listElementType, maxListSize));
+    List<SszListSchema<TElement, ?>> types =
+        generateTypesWithHints(SszListSchema.create(listElementType, maxListSize));
 
     RewindingSupplier<TElement> rewindingSupplier = new RewindingSupplier<>(listElementsFactory);
     ArrayList<SszList<TElement>> resultsToCompare = new ArrayList<>();
     testList(types.get(0), rewindingSupplier, resultsToCompare::add);
 
     for (int i = 1; i < types.size(); i++) {
-      SszListSchema<TElement> type = types.get(i);
+      SszListSchema<TElement, ?> type = types.get(i);
       rewindingSupplier.rewind();
       ArrayDeque<SszList<TElement>> resQueue = new ArrayDeque<>(resultsToCompare);
       testList(
@@ -225,7 +225,7 @@ public class ListViewHintsTest {
   }
 
   <TElement extends SszData> void testList(
-      SszListSchema<TElement> type,
+      SszListSchema<TElement, ?> type,
       Supplier<TElement> listElementsFactory,
       Consumer<SszList<TElement>> results) {
